+++ internal/http/www/server.go
@@ -156,8 +156,8 @@ func (s *Server) registerHandlers() {
 
 	// Domain handlers
 	controller.RegisterSessionHandlers(s.router, s.sessionService, s.userService)
-	controller.RegisterAdminHandlers(s.router, s.userService)
-	controller.RegisterAccountHandlers(s.router, s.feedService, s.userService)
+	controller.RegisterAdminHandlers(s.router, s.sessionService, s.userService)
+	controller.RegisterAccountHandlers(s.router, s.feedService, s.sessionService, s.userService)
 	controller.RegisterBookmarkHandlers(s.router, s.publicURL, s.bookmarkService, s.bookmarkExportingService, s.bookmarkImportingService, s.bookmarkQueryingService, s.userService)
 	controller.RegisterFeedHandlers(s.router, s.feedService, s.feedExportingService, s.feedImportingService, s.feedQueryingService, s.userService)