+++ docs/src/changelog.md
@@ -9,6 +9,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ### Security
 - Update Chi HTTP middleware to read the remote client IP address from a header set by a reverse proxy
 - Prevent the HTTP client used to fetch syndication feeds from performing requests to unroutable or blocked IP address ranges.
+- Fully invalidate user sessions on logout and password change
 
 ### Fixed
 #### Packaging and automation